requirements.txt
certifi==2025.8.3
protobuf==6.32.0
websocket-client==1.8.0
